Updated scripts

+++ scripts.py
@@ -1,45 +1,53 @@
-from django.core.exceptions import ObjectDoesNotExist, MultipleObjectsReturned
-from datacenter.models import (Chastisement, Commendation, Lesson, Mark, Schoolkid, Subject, Teacher)
+from datacenter.models import (Chastisement, Commendation, Lesson, Mark, Schoolkid)
 import random
 import logging
+import sys
 
+logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-def fix_marks(schoolkid, child_name):
-    child = schoolkid.objects.filter(full_name__contains=child_name)
-    bad_marks = Mark.objects.filter(schoolkid=child[0], points__lt=4)
-    for bad_mark in bad_marks:
-        bad_mark.points = 5
-        bad_mark.save()
+def get_child_by_name(child_name):
+    try:
+        return Schoolkid.objects.get(full_name__contains=child_name)
+    except Schoolkid.DoesNotExist:
+        logger.error(f'Произошла ошибка: Ученика с указанным именем не найдено.')
+        sys.exit()
+    except Schoolkid.MultipleObjectsReturned:
+        logger.error(f'Произошла ошибка: Получено несколько учеников с этим именем.')
+        sys.exit()
 
 
-def remove_chastisements(schoolkid, child_name):
-    child = schoolkid.objects.filter(full_name__contains=child_name)
-    chastisements = Chastisement.objects.filter(schoolkid=child[0])
+def fix_marks(child_name):
+    child = get_child_by_name(child_name)
+    bad_marks = Mark.objects.filter(schoolkid=child, points__lt=4)
+    if not bad_marks:
+        return logger.info('У данного ученика нет плохих оценок')
+    bad_marks.update(points=random.randint(4, 5))
+
+
+def remove_chastisements(child_name):
+    child = get_child_by_name(child_name)
+    chastisements = Chastisement.objects.filter(schoolkid=child)
+    if not chastisements:
+        return logger.info('У данного ученика нет замечаний')
     chastisements.delete()
 
 
+
 def create_commendation(child_name, lesson_title):
-    try:
-        year_of_study = 6
-        group_letter = 'А'
-        praises = ['Молодец', 'Так держать!', 'У тебя всё получается!', 'Успех близко', 'Мне приятно от твоих успехов!']
-
-        children = Schoolkid.objects.filter(year_of_study=year_of_study, group_letter=group_letter, full_name__contains=child_name)
-
-        if not children.exists():
-            raise ObjectDoesNotExist('Данный пользователь не найден')
-
-        if children.count() > 1:
-            raise MultipleObjectsReturned('Возвращено несколько значений')
-
-        subject = Subject.objects.filter(title=lesson_title, year_of_study=year_of_study).first()
-        lesson = Lesson.objects.filter(year_of_study=year_of_study, group_letter=group_letter, subject=subject).first()
-        teacher = Teacher.objects.filter(full_name=lesson.teacher).first()
-        Commendation.objects.create(schoolkid=children[0], subject=subject, teacher=teacher, text=random.choice(praises), created=lesson.date)
-
-    except ObjectDoesNotExist as e:
-        logger.warning(f'Произошла ошибка: {e}')
-    except MultipleObjectsReturned as e:
-        logger.warning(f'Произошла ошибка: {e}')
+    compliments = ['Молодец', 'Так держать!', 'У тебя всё получается!', 'Успех близко', 'Мне приятно от твоих успехов!']
+    child = get_child_by_name(child_name)
+    lessons = Lesson.objects.filter(
+        year_of_study=child.year_of_study, 
+        group_letter=child.group_letter, 
+        subject__title=lesson_title
+        )
+    lesson_of_subject = lessons.order_by('-date').first()
+    Commendation.objects.create(
+        schoolkid=child, 
+        subject=lesson_of_subject.subject, 
+        teacher=lesson_of_subject.teacher, 
+        text=random.choice(compliments), 
+        created=lesson_of_subject.date
+    )
